Modificaciones
1) Creación de comentarios para los métodos de las clases servicios.

2) Se modifica método de actualización de servicio GeneroServicesImpl.

3) Modifica métodos de eliminación de la clase servicios CategoriaServicesImpl.

4) Modifica métodos de eliminació de la clase servicios EditorialServicesImpl.

5) Modifica métodos de eliminación de la clase PaisServicesImpl.

+++ src/main/java/com/complejo/educacional/luis/durand/durand/services/CategoriaServicesImpl.java
@@ -1,4 +1,5 @@
 package com.complejo.educacional.luis.durand.durand.services;
+//Import necesarios para la clase.
 
 import com.complejo.educacional.luis.durand.durand.dto.categoria.CategoriaDTORequest;
 import com.complejo.educacional.luis.durand.durand.dto.categoria.CategoriaDTOResponse;
@@ -28,11 +29,12 @@ public class CategoriaServicesImpl implements ICategoriaServices {
     private ObjectMapper objectMapper;
 
     /**
-     * @param categoriaDTORequest
-     * @return categoriaDTORequest
-     * @throws Exception
+     * Método que guarda una nueva categoría con la información proporcionada en categoriaDTORequest.
+     *
+     * @param categoriaDTORequest La información de la nueva categoría a guardar.
+     * @return Un objeto CategoriaDTORequest con la información de la categoría guardado.
+     * @throws Exception Si ocurre un error durante el proceso de guardado de la categoría.
      */
-
     @Override
     @Transactional(readOnly = false)
     public CategoriaDTORequest saveCategoria(CategoriaDTORequest categoriaDTORequest) throws Exception {
@@ -41,9 +43,7 @@ public class CategoriaServicesImpl implements ICategoriaServices {
             Categoria categoriaCreate = new Categoria(
                     null,
                     categoriaDTORequest.getNombre_categoria(),
-                    categoriaDTORequest.getDescripcion_categoria(),
-                    categoriaDTORequest.getCreateAt(),
-                    categoriaDTORequest.getCreateAt()
+                    categoriaDTORequest.getDescripcion_categoria()
             );
             log.info("¡Creación de Categoría!");
             iCategoriaRepository.save(categoriaCreate);
@@ -56,10 +56,12 @@ public class CategoriaServicesImpl implements ICategoriaServices {
     }
 
     /**
-     * @param id
-     * @param categoriaRequestUpdate
-     * @return categoriaDTOResponse
-     * @throws Exception
+     * Método que actualiza una categoría según su ID con la información proporcionada en categoriaRequestUpdate.
+     *
+     * @param id                     El ID de la categoría a actualizar.
+     * @param categoriaRequestUpdate La información actualizada de la categoría.
+     * @return Un objeto CategoriaRequestUpdate con la información actualizada de la categoría.
+     * @throws Exception Si ocurre un error durante el proceso de actualización de la categoría.
      */
     @Override
     @Transactional(readOnly = false)
@@ -69,9 +71,7 @@ public class CategoriaServicesImpl implements ICategoriaServices {
         Categoria categoriaUpdate = new Categoria(
                 categoriaRequestUpdate.getId_categoria(),
                 categoriaRequestUpdate.getNombre_categoria(),
-                categoriaRequestUpdate.getDescripcion_categoria(),
-                categoriaRequestUpdate.getCreateAt(),
-                categoriaRequestUpdate.getUdpdateAt()
+                categoriaRequestUpdate.getDescripcion_categoria()
         );
         try {
             optionalCategoria = iCategoriaRepository.findById(id);
@@ -84,9 +84,7 @@ public class CategoriaServicesImpl implements ICategoriaServices {
             categoriaDTOResponse = new CategoriaDTOResponse(
                     categoriaUpdate.getId_categoria(),
                     categoriaUpdate.getNombre_categoria(),
-                    categoriaUpdate.getDescripcion_categoria(),
-                    categoriaUpdate.getCreateAt(),
-                    categoriaUpdate.getUdpdateAt()
+                    categoriaUpdate.getDescripcion_categoria()
             );
             return categoriaDTOResponse;
         } catch (Exception e) {
@@ -96,9 +94,11 @@ public class CategoriaServicesImpl implements ICategoriaServices {
     }
 
     /**
-     * @param sort
-     * @return categoriaDTOResponses
-     * @throws Exception
+     * Método que recupera una lista de objetos CategoriaDTOResponse ordenados según el criterio especificado.
+     *
+     * @param sort El criterio de ordenación.
+     * @return Una lista de objetos CategoriaDTOResponse ordenados según el criterio especificado.
+     * @throws Exception Si ocurre un error al recuperar la lista de las categorías ordenadas.
      */
     @Override
     @Transactional(readOnly = true)
@@ -109,9 +109,7 @@ public class CategoriaServicesImpl implements ICategoriaServices {
                     new CategoriaDTOResponse(
                             categoria.getId_categoria(),
                             categoria.getNombre_categoria(),
-                            categoria.getDescripcion_categoria(),
-                            categoria.getCreateAt(),
-                            categoria.getUdpdateAt()
+                            categoria.getDescripcion_categoria()
                     )
             );
         }
@@ -119,10 +117,11 @@ public class CategoriaServicesImpl implements ICategoriaServices {
     }
 
     /**
+     * Método que recupera una página de objetos CategoriaTOResponse utilizando paginación y ordenación.
      *
-     * @param pageable
-     * @return categoriaDTOResponses
-     * @throws Exception
+     * @param pageable La información de paginación y ordenación.
+     * @return Una página de objetos CategoriaDTOResponse.
+     * @throws Exception Si ocurre un error al recuperar la página de las categorías.
      */
     @Override
     @Transactional(readOnly = true)
@@ -133,9 +132,7 @@ public class CategoriaServicesImpl implements ICategoriaServices {
                     new CategoriaDTOResponse(
                             categoria.getId_categoria(),
                             categoria.getNombre_categoria(),
-                            categoria.getDescripcion_categoria(),
-                            categoria.getCreateAt(),
-                            categoria.getUdpdateAt()
+                            categoria.getDescripcion_categoria()
                     )
             );
         }
@@ -143,10 +140,11 @@ public class CategoriaServicesImpl implements ICategoriaServices {
     }
 
     /**
+     * Método que busca una categoría por su ID y devuelve su información en un objeto CategoriaDTOResponse.
      *
-     * @param id
-     * @return CategoriaDTOResponse
-     * @throws Exception
+     * @param id El ID de la categoría a buscar.
+     * @return Un objeto CategoriaDTOResponse con la información de la categoría.
+     * @throws Exception Si la categoría no se encuentra o si ocurre un error durante la búsqueda.
      */
     @Override
     @Transactional(readOnly = true)
@@ -158,28 +156,31 @@ public class CategoriaServicesImpl implements ICategoriaServices {
         return new CategoriaDTOResponse(
                 categoriaId.getId_categoria(),
                 categoriaId.getNombre_categoria(),
-                categoriaId.getDescripcion_categoria(),
-                categoriaId.getCreateAt(),
-                categoriaId.getUdpdateAt()
+                categoriaId.getDescripcion_categoria()
         );
     }
 
     /**
+     * Método para eliminar una categoría según su ID.
      *
-     * @param id
-     * @return eliminarCategoria
-     * @throws Exception
+     * @param id El ID de la categoría a eliminar.
+     * @return true si se elimina de la categoría con éxito.
+     * @throws Exception Si ocurre un error durante el proceso de eliminación de la categoría.
      */
     @Override
     @Transactional(readOnly = false)
-    public Object deleteCategoriaById(Long id) throws Exception {
-        Object eliminarCategoria= null;
+    public boolean deleteCategoriaById(Long id) throws Exception {
         try {
-            log.info("¡Eliminar Categoría");
-            iCategoriaRepository.deleteById(id);
-        }catch (Exception e){
-            log.error("¡Falló la eliminación de la Categoría"+e.getCause().toString());
+            if (iCategoriaRepository.existsById(id)) {
+                log.info("Eliminar Categoría con ID: " + id);
+                iCategoriaRepository.deleteById(id);
+                return true;
+            } else {
+                log.error("¡No existe el ID del Categoría!");
+                throw new Exception("¡No existe el ID del Categoría!");
+            }
+        } catch (Exception e) {
+            log.error("¡Falló la eliminación de la Categoría" + e.getCause().toString());
         }
-        return eliminarCategoria;
     }
 }
